src/ai/requirements.txt
pandas
numpy
joblib
python-dateutil
scikit-learn
